Move grouping & enrollment logic to services
Extract assignment grouping, class enrollment, and review dumping logic into dedicated service modules and helper controllers. Added assignment_grouping_controller, class_enrollment_controller, controllers/helpers/assignment_access, and services: assignment_grouping_service, class_enrollment_service, review_service; updated imports and usages in assignment_controller, class_controller, review_controller and api.__init__ to register new blueprints. Replaced in-controller CSV parsing, group management, and review serialization with service calls (including clear_assignment_groups, replace_group_members, build_grouping_student_payload, csv_to_list, enroll_students_in_course, dump_review_with_markable_criteria, dump_received_review_anonymized, split_reviews_by_type) to improve separation of concerns and testability.
