+++ flask_backend/api/controllers/helpers/__init__.py
@@ -1,0 +1,3 @@
+from .assignment_access import get_authenticated_user, get_teacher_managed_assignment
+
+__all__ = ["get_authenticated_user", "get_teacher_managed_assignment"]
